requirements.txt
annotated-types==0.7.0
anyio==4.6.0
cassandra-driver==3.29.2
click==8.1.7
fastapi==0.115.0
geomet==0.2.1.post1
h11==0.14.0
idna==3.10
pydantic==2.9.2
pydantic_core==2.23.4
PyYAML==6.0.2
scylla-driver==3.26.9
six==1.16.0
sniffio==1.3.1
starlette==0.38.6
typing_extensions==4.12.2
uvicorn==0.30.6
